requirements.txt
absl-py==1.0.0
alabaster==0.7.12
asn1crypto==0.24.0
astunparse==1.6.3
attrs==17.4.0
Automat==0.6.0
Babel==2.9.1
black==21.11b1
bleach==4.1.0
bob==10.0.0
bob.blitz==2.0.23
#bob.core==2.2.7
bob.devtools==4.1.0
bob.extension==5.0.1
#bob.io.base==3.0.10
#bob.ip.gabor==2.0.18
#bob.sp==2.0.17
BobBuildTool==0.19.2
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contextlib2==21.6.0
cryptography==2.1.4
cycler==0.11.0
Cython==0.29.25
distro-info===0.18ubuntu0.18.04.1
docformatter==1.4
docutils==0.17.1
flatbuffers==2.0
fonttools==4.28.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.24
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.18.2
grpcio==1.42.0
h5py==3.6.0
httplib2==0.9.2
hyperlink==17.3.1
idna==3.3
imageio==2.12.0
imagesize==1.3.0
importlib-metadata==4.8.2
incremental==16.10.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
language-selector==0.1
libclang==12.0.0
llvmlite==0.37.0
lxml==4.6.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
mypy-extensions==0.4.3
natsort==8.0.0
netifaces==0.10.4
networkx==2.6.3
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
PAM==0.4.2
pandas==1.3.4
pathspec==0.9.0
pexpect==4.2.1
Pillow==8.4.0
pkginfo==1.8.1
platformdirs==2.4.0
protobuf==3.19.1
psutil==5.8.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycurl==7.43.0.1
Pygments==2.10.0
PyGObject==3.26.1
pyOpenSSL==17.5.0
pyparsing==3.0.6
pyserial==3.4
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
python-debian==0.1.32
python-gitlab==2.10.1
python-magic==0.4.24
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
readme-renderer==30.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rsa==4.8
schema==0.7.4
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.3
service-identity==16.0.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
snowballstemmer==2.2.0
sos==4.1
Sphinx==4.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
ssh-import-id==5.7
systemd-python==234
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tomli==1.2.2
torch==1.10.0
tqdm==4.62.3
twine==2.0.0
Twisted==17.9.0
typing_extensions==4.0.0
ubuntu-advantage-tools==27.4
ufw==0.36
unattended-upgrades==0.1
untokenize==0.1.1
urllib3==1.26.7
webdavclient3==3.14.6
webencodings==0.5.1
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
zope.interface==4.3.2
